requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=2.0.0
beautifulsoup4>=4.9.3
requests>=2.26.0
python-dotenv>=0.19.0
email-validator>=1.1.3
python-imap>=1.0.0
aiohttp>=3.8.1
python-dateutil>=2.8.2
lxml>=4.9.0 